remote_agents/cosmos_query_agent/pyproject.toml
[project]
name = "cosmos-query-agent"
version = "0.1.0"
description = "A2A Cosmos Query Agent using Azure Cosmos DB MCP Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.0",
    "mcp>=1.9.4",
    "azure-ai-agents>=1.1.0b1",
    "semantic-kernel>=1.33.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "click>=8.0.0",
    "uvicorn>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cosmos_query_agent"]
